This was causing startup to fail on subsequent container restarts where the file had already been moved.

+++ config-opinions/cf-v217/prevent_apparmor_disable.sh
@@ -3,4 +3,7 @@
 # runner attempts to disable apparmor which ends up disabling it on the host
 # preventing other containers from starting. It checks for this init script
 # before doing so, so we move it to a different location.
-mv /etc/init.d/apparmor /etc/init.d/apparmor_host
+
+if [[ -f /etc/init.d/apparmor ]] ; then
+    mv /etc/init.d/apparmor /etc/init.d/apparmor_host
+fi
